correctly specify UTF-8 encoding and load jars from the lockfile in YAML format

+++ lib/ruby/pathname.rb
@@ -1,7 +1,7 @@
 class Pathname
   def touch(path)
     file = self + path
-    file.open("w+") { nil }
+    file.open("w+") { nil } unless file.exist?
     file
   end
